Separate capturing/reporting and 'caplog' recording/inspection features
This makes a distinction between two features / use cases:

  1. Inspection: record and check logs emitted by the code being tested
     using the `caplog` fixture

  2. Reporting: show all logs in case of a failure

Implemented through adding a dedicated log handler for the recording.
As a bonus, simplifies the capturing logic (no more awkward
item.catch_log_handler).

Also move LogCaptureHandler into fixture.py.
