+++ pytest_catchlog/fixture.py
@@ -10,16 +10,39 @@ import py
 from pytest_catchlog.common import catching_logs, logging_at_level
 
 
+BASIC_FORMATTER = logging.Formatter(logging.BASIC_FORMAT)
+
+
+class LogCaptureHandler(logging.StreamHandler):
+    """A logging handler that stores log records and the log text."""
+
+    def __init__(self):
+        """Creates a new log handler."""
+
+        logging.StreamHandler.__init__(self)
+        self.stream = py.io.TextIO()
+        self.records = []
+
+    def close(self):
+        """Close this log handler and its underlying stream."""
+
+        logging.StreamHandler.close(self)
+        self.stream.close()
+
+    def emit(self, record):
+        """Keep the log records in a list in addition to the log text."""
+
+        self.records.append(record)
+        logging.StreamHandler.emit(self, record)
+
+
 class LogCaptureFixture(object):
     """Provides access and control of log capturing."""
 
-    @property
-    def handler(self):
-        return self._item.catch_log_handler
-
-    def __init__(self, item):
+    def __init__(self, handler):
         """Creates a new funcarg."""
-        self._item = item
+        super(LogCaptureFixture, self).__init__()
+        self.handler = handler
 
     @property
     def text(self):
@@ -104,6 +127,11 @@ class CallableStr(CallablePropertyMixin, py.builtin.text):
 class CompatLogCaptureFixture(LogCaptureFixture):
     """Backward compatibility with pytest-capturelog."""
 
+    def __init__(self, handler, item):
+        """Creates a new funcarg."""
+        super(CompatLogCaptureFixture, self).__init__(handler)
+        self._item = item
+
     def _warn_compat(self, old, new):
         self._item.warn(code='L1',
                         message=("{0} is deprecated, use {1} instead"
@@ -132,7 +160,7 @@ class CompatLogCaptureFixture(LogCaptureFixture):
         return self.at_level(level, logger)
 
 
-@pytest.fixture
+@pytest.yield_fixture
 def caplog(request):
     """Access and control log capturing.
 
@@ -142,6 +170,8 @@ def caplog(request):
     * caplog.records()       -> list of logging.LogRecord instances
     * caplog.record_tuples() -> list of (logger_name, level, message) tuples
     """
-    return CompatLogCaptureFixture(request.node)
+    with catching_logs(LogCaptureHandler(),
+                       formatter=BASIC_FORMATTER) as handler:
+        yield CompatLogCaptureFixture(handler, item=request.node)
 
 capturelog = caplog
